chore(client): move regression tests to their own directory
