--- packages/client/tests/functional/21136-extensions-mocking-library/_matrix.ts
@@ -1,10 +1,0 @@
-import { defineMatrix } from '../_utils/defineMatrix'
-import { Providers } from '../_utils/providers'
-
-export default defineMatrix(() => [
-  [
-    {
-      provider: Providers.SQLITE,
-    },
-  ],
-])
