--- packages/client/tests/functional/24835-omit-error/_matrix.ts
@@ -1,4 +1,0 @@
-import { defineMatrix } from '../_utils/defineMatrix'
-import { allProviders } from '../_utils/providers'
-
-export default defineMatrix(() => [allProviders])
