+++ packages/client/tests/functional/issues/17797-no-env-error-init/prisma/_schema.ts
@@ -1,4 +1,4 @@
-import { idForProvider } from '../../_utils/idForProvider'
+import { idForProvider } from '../../../_utils/idForProvider'
 import testMatrix from '../_matrix'
 
 export default testMatrix.setupSchema(({ provider }) => {
@@ -9,12 +9,11 @@ export default testMatrix.setupSchema(({ provider }) => {
   
   datasource db {
     provider = "${provider}"
-    url      = env("DATABASE_URI_${provider}")
+    url      = env("INVALID_DATABASE_URI")
   }
   
   model User {
     id ${idForProvider(provider)}
-    name String
   }
   `
 })
