+++ packages/client/tests/functional/issues/17797-no-env-error-init/tests.ts
@@ -1,7 +1,8 @@
-import { NewPrismaClient } from '../_utils/types'
-import testMatrix from './_matrix'
 // @ts-ignore
-import { PrismaClient } from './node_modules/@prisma/client'
+import { PrismaClient } from '@prisma/client'
+
+import { NewPrismaClient } from '../../_utils/types'
+import testMatrix from './_matrix'
 
 declare const newPrismaClient: NewPrismaClient<typeof PrismaClient>
