+++ packages/client/tests/functional/issues/17810-direct-url-missing/tests.ts
@@ -1,7 +1,8 @@
-import { NewPrismaClient } from '../_utils/types'
-import testMatrix from './_matrix'
 // @ts-ignore
-import type { PrismaClient } from './node_modules/@prisma/client'
+import type { PrismaClient } from '@prisma/client'
+
+import { NewPrismaClient } from '../../_utils/types'
+import testMatrix from './_matrix'
 
 let prisma: PrismaClient
 declare let newPrismaClient: NewPrismaClient<typeof PrismaClient>
