+++ packages/client/tests/functional/issues/20422-cannot-assign-type-to-itself/prisma/_schema.ts
@@ -1,4 +1,4 @@
-import { idForProvider } from '../../_utils/idForProvider'
+import { idForProvider } from '../../../_utils/idForProvider'
 import testMatrix from '../_matrix'
 
 export default testMatrix.setupSchema(({ provider }) => {
@@ -14,6 +14,7 @@ export default testMatrix.setupSchema(({ provider }) => {
   
   model User {
     id ${idForProvider(provider)}
+    name String
   }
   `
 })
