+++ packages/client/tests/functional/issues/20422-cannot-assign-type-to-itself/tests.ts
@@ -1,7 +1,8 @@
 /* eslint-disable @typescript-eslint/no-redundant-type-constituents */
-import testMatrix from './_matrix'
 // @ts-ignore
-import type { Prisma as PrismaNamespace, PrismaClient } from './node_modules/@prisma/client'
+import type { Prisma as PrismaNamespace, PrismaClient } from '@prisma/client'
+
+import testMatrix from './_matrix'
 
 testMatrix.setupTestSuite(
   () => {
