+++ packages/client/tests/functional/issues/21136-extensions-mocking-library/_matrix.ts
@@ -1,0 +1,10 @@
+import { defineMatrix } from '../../_utils/defineMatrix'
+import { Providers } from '../../_utils/providers'
+
+export default defineMatrix(() => [
+  [
+    {
+      provider: Providers.SQLITE,
+    },
+  ],
+])
