+++ packages/client/tests/functional/issues/21136-extensions-mocking-library/prisma/_schema.ts
@@ -1,4 +1,4 @@
-import { idForProvider } from '../../_utils/idForProvider'
+import { idForProvider } from '../../../_utils/idForProvider'
 import testMatrix from '../_matrix'
 
 export default testMatrix.setupSchema(({ provider }) => {
@@ -9,7 +9,7 @@ export default testMatrix.setupSchema(({ provider }) => {
   
   datasource db {
     provider = "${provider}"
-    url      = env("INVALID_DATABASE_URI")
+    url      = env("DATABASE_URI_${provider}")
   }
   
   model User {
