+++ packages/client/tests/functional/issues/24835-omit-error/prisma/_schema.ts
@@ -1,4 +1,4 @@
-import { idForProvider } from '../../_utils/idForProvider'
+import { idForProvider } from '../../../_utils/idForProvider'
 import testMatrix from '../_matrix'
 
 export default testMatrix.setupSchema(({ provider }) => {
