+++ packages/client/tests/functional/issues/24835-omit-error/test.ts
@@ -1,6 +1,7 @@
-import testMatrix from './_matrix'
 // @ts-ignore
-import type { PrismaClient } from './node_modules/@prisma/client'
+import type { PrismaClient } from '@prisma/client'
+
+import testMatrix from './_matrix'
 
 declare let prisma: PrismaClient
